scaled subOp with rescaled scalar input (#29)
For both scaled add and sub op with scalar cases, we fold the scalar rescale op chain into a
single constant. Then we use standard add op kernel with bias parameter for computation.

In order to compute the constant value we use the output scale and shift factor (in case the
output is rescaled).

This PR got input from Maurizio.

Co-authored-by: Xiaomei.Shang <Xiaomei.Shang@synaptics.com>