Move back the tile-and-fuse marking pass after the linalg optimization for torq pass (#397)
The tile-and-fuse marking pass was "abused" in order to mark rescale operations in order to not broadcast in those
cases. This rescale check verification is now done directly without help.

This is the correct thing to do because we want the tile-and-fuse marking to be done after the linalg transformation
is completely done.