Add test cases to main repo (#370)
A new test infrastructure is added that can be used both for in-tree
and out of tree tests.

The infrastructure allows to cache intermediate data such as
computed reference results and MLIR files.

A number of linalg and tosa based test files are added.

A number of keras based tests are added.