fix(torch): Switch torch to linalg pipeline to use single dispatch (#401)
Stock IREE would lower torch IR to linalg using multiple dispatches; with this commit we switch to as single
dispatch. This is inline with what we already do for TOSA input.

Downside is that we don't benefit from IREE optimization when not using its pipeline, this will be revisited
when we are truly ready to use heterogeneous inference at a higher level.