Restructure testing and add cross-test caching (#406)
Instead of using class-based tests and inheritance we use fixtures
that are more idiomatic to pytest and easier to understand.

In order to speed up testing the commit add caching for fixtures
that are re-used across tests. Previously we were re-computing
each piece of data for each test individually, this caused some
data to be re-computed multiple times.

Now the code adds support for re-using fixtures across tests and
pytest sessions. This means that running test for different targets
can re-use the same pre-computed data. This can cut down the testing
time on FPGA and SOC significantly (most of the data can be re-used
from cmodel testing). Same applies for cross target tests. Same
reference results can be re-used for different targets.

In order to avoid mistakes the code adds support for versioning
the fixtures. This makes sure that if a test requests a different
version of the fixture (which depends possibly on the parameters
of the fixture and external files), the feature is automatically
recomputed.

The commit also adds support for reading tests from extras/ if present.