Add support for testing multiple chips
The commit adds support to test for multiple chips.

The chips can either be defined with a YAML file in
extras/chips or by name (as used in the --torq-hw option).

By default tests use the SL2610 chip.

The commit also fixes cache invalidation when passing extra
command line options to torq compiler and runtime and
adds a new fixture versioned_generated_directory_fixture that
is used to properly store ancillary information of model
builds.