Add support for supporting multiple CSS configuration (#471)
Currently we hardcoded only support for kelvin/coral and qemu.

If we want to experiment with different revisions of the CSS
core we need to be able to compile in the compiler bootstraps
suitable for them.

This commit uses the new infra to cleanly divide the binaries
used to build for qemu and for kelvin.

Private configuration can be added in extras/css_configs.

The commit also adds code to support fp instructions in
dispatches.