Add support for soft-float sqrt and other math fun (#557)
The soft-float implementations of the math functions use errno. By default this is in thread local storage.

The commit adds a new build of libc and compiler_rt  that doesn't use thread local storage (not available
on CSS) at all.

The build also adds got and sbss to the linker script used for errno so that now calls to sqrt and similar
functions work out of the box.

The commit then disables the automatic fallback.